Added support for a minLength of 0, added all options for items count and added showHintOnFocus option
Added support for a minLength of 0 that displays the hints on focus and when user clears the field. The showHintsOnFocus can be set to true so hints are shown whenever user enters the field (even if minLength is grater than 0). You can also set the count of items to 'all' if you do not want to limit the number of items shown. This is always the case when minLength == 0 and no value is enter (all the hints are shown).