Handle tabbing into a field with showHintOnFocus
As noted in
https://github.com/bassjobsen/Bootstrap-3-Typeahead/issues/107, when
`showHintOnFocus` is not `false` and a user tabs into an auto-select field,
the `keyUp` from the tab press causes the first item in the list to be
selected.

This is rarely what is wanted, so this commit changes the `keyup`
handler to ignore tabs when the list is shown and `showHintOnFocus` is
set.

Note that this does mean that pressing the tab key when
`showHintOnFocus` is set will never select the highlighted item if any, whereas
when `showHintOnFocus` is false it will.